+++ CHANGELOG.md
@@ -68,6 +68,8 @@
 
 ### v3.3.10 / 2020-07-xx
 
+* [599](https://github.com/Vanessa219/vditor/issues/599) Ordered list(minor bug) `修复缺陷`
+
 ### v3.3.9 / 2020-07-18
 
 * [591](https://github.com/Vanessa219/vditor/issues/591) 粘贴有可能不渲染本行内容 `改进功能`